+++ frontend_web/frontend_web/src/components/shared/DeleteConfirmationDialog.jsx
@@ -1,0 +1,32 @@
+import React from 'react';
+
+function DeleteConfirmationDialog({ isOpen, onClose, onDelete }) {
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 z-50 overflow-y-auto bg-black bg-opacity-50 flex items-center justify-center">
+      <div className="bg-white rounded-lg max-w-md w-full p-6 shadow-xl">
+        <h3 className="text-lg font-medium text-gray-900 mb-4">Confirm Deletion</h3>
+        <p className="text-gray-600 mb-6">
+          Are you sure you want to delete this address? This action cannot be undone.
+        </p>
+        <div className="flex justify-end space-x-3">
+          <button
+            onClick={onClose}
+            className="border border-gray-300 text-gray-700 px-4 py-2 rounded hover:bg-gray-50 transition"
+          >
+            Cancel
+          </button>
+          <button
+            onClick={onDelete}
+            className="bg-red-600 text-white px-4 py-2 rounded hover:bg-red-700 transition"
+          >
+            Delete
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default DeleteConfirmationDialog;
